Separated Auto Configuration from Starter
Fixes #208
