+++ tracer-spring-boot-autoconfigure/src/main/java/org/zalando/tracer/autoconfigure/TracerAutoConfiguration.java
@@ -1,4 +1,4 @@
-package org.zalando.tracer.spring;
+package org.zalando.tracer.autoconfigure;
 
 import io.opentracing.Tracer;
 import io.opentracing.contrib.api.tracer.APIExtensionsTracer;
