+++ tracer-spring-boot-autoconfigure/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,1 @@
+org.springframework.boot.autoconfigure.EnableAutoConfiguration = org.zalando.tracer.autoconfigure.TracerAutoConfiguration
