--- tracer-spring-boot-starter/src/main/resources/META-INF/spring.factories
@@ -1,1 +1,0 @@
-org.springframework.boot.autoconfigure.EnableAutoConfiguration = org.zalando.tracer.spring.TracerAutoConfiguration
